@@ -4,8 +4,10 @@
 package inspect
 
 import (
+	"cmp"
 	"context"
 	"fmt"
+	"slices"
 	"sort"
 
 	"github.com/google/cel-go/common/ast"
@@ -17,7 +19,10 @@ import (
 	"github.com/cerbos/cerbos/internal/namer"
 	"github.com/cerbos/cerbos/internal/policy"
 	"github.com/cerbos/cerbos/internal/storage"
+	"github.com/cerbos/cerbos/internal/util"
 )
+
+type loadPolicyFn func(ctx context.Context, policyKey ...string) ([]*policy.Wrapper, error)
 
 func Policies() *Policy {
 	return &Policy{
@@ -49,41 +54,24 @@ func (pol *Policy) Inspect(p *policyv1.Policy) error {
 	}
 
 	policyID := namer.PolicyKey(p)
-	localVariables := policy.ListVariables(p)
-	if _, ok := p.PolicyType.(*policyv1.Policy_ExportVariables); ok {
-		sortedLocalVariables := make([]*responsev1.InspectPoliciesResponse_Variable, 0, len(localVariables))
-		for _, lv := range localVariables {
-			sortedLocalVariables = append(sortedLocalVariables, lv)
-		}
-		sort.Slice(sortedLocalVariables, func(i, j int) bool {
-			return sortedLocalVariables[i].Name < sortedLocalVariables[j].Name
-		})
+	pol.derivedRolesImports[policyID], pol.variableImports[policyID] = pol.listImports(p)
 
-		if len(localVariables) > 0 {
-			pol.results[policyID] = &responsev1.InspectPoliciesResponse_Result{
-				Variables: sortedLocalVariables,
-				PolicyId:  storeIdentifier,
-			}
-		}
-
-		return nil
-	}
-
-	referencedDerivedRoles, referencedVariables, err := pol.inspectDefinitionsAndRules(p)
+	attributes, err := pol.listReferencedAttributes(p)
 	if err != nil {
-		return fmt.Errorf("failed to inspect definitions and rules of the policy %s: %w", policyID, err)
+		return fmt.Errorf("failed to list referenced attributes in the policy %s: %w", policyID, err)
 	}
 
 	var derivedRoles []*responsev1.InspectPoliciesResponse_DerivedRole
-
+	//nolint:nestif
 	if drp := p.GetDerivedRoles(); drp != nil {
 		derivedRoles = policy.ListExportedDerivedRoles(drp)
 		if len(derivedRoles) > 0 {
-			sort.Slice(derivedRoles, func(i, j int) bool {
-				return derivedRoles[i].Name < derivedRoles[j].Name
+			slices.SortFunc(derivedRoles, func(a, b *responsev1.InspectPoliciesResponse_DerivedRole) int {
+				return cmp.Compare(a.GetName(), b.GetName())
 			})
 		}
-	} else {
+	} else if rp := p.GetResourcePolicy(); rp != nil {
+		referencedDerivedRoles := pol.listReferencedDerivedRoles(rp)
 		for referenced := range referencedDerivedRoles {
 			if toResolve, exists := pol.derivedRolesToResolve[policyID]; !exists {
 				pol.derivedRolesToResolve[policyID] = map[string]bool{referenced: false}
@@ -93,8 +81,12 @@ func (pol *Policy) Inspect(p *policyv1.Policy) error {
 		}
 	}
 
-	pol.derivedRolesImports[policyID], pol.variableImports[policyID] = pol.listImports(p)
+	referencedVariables, err := pol.listReferencedVariables(p)
+	if err != nil {
+		return fmt.Errorf("failed to list referenced variables in the policy %s: %w", policyID, err)
+	}
 
+	localVariables := policy.ListVariables(p)
 	for referenced := range referencedVariables {
 		if v, ok := localVariables[referenced]; ok {
 			v.Used = true
@@ -113,28 +105,36 @@ func (pol *Policy) Inspect(p *policyv1.Policy) error {
 	}
 
 	// sort variables if there is nothing to resolve since we are not going to modify variables in the future.
-	if len(pol.variablesToResolve[policyID]) > 0 {
-		sort.Slice(variables, func(i, j int) bool {
-			return variables[i].Name < variables[j].Name
+	if len(pol.variablesToResolve[policyID]) == 0 {
+		slices.SortFunc(variables, func(a, b *responsev1.InspectPoliciesResponse_Variable) int {
+			return cmp.Compare(a.GetName(), b.GetName())
 		})
 	}
 
-	a := policy.ListActions(p)
-	sort.Strings(a)
+	actions := policy.ListActions(p)
+	sort.Strings(actions)
 	pol.results[policyID] = &responsev1.InspectPoliciesResponse_Result{
-		Actions:      a,
+		Actions:      actions,
+		Attributes:   attributes,
 		DerivedRoles: derivedRoles,
-		Variables:    variables,
 		PolicyId:     storeIdentifier,
+		Variables:    variables,
 	}
 
 	return nil
 }
 
-type loadPolicyFn func(ctx context.Context, policyKey ...string) ([]*policy.Wrapper, error)
-
 // Results returns the final inspection results.
 func (pol *Policy) Results(ctx context.Context, loadPolicy loadPolicyFn) (map[string]*responsev1.InspectPoliciesResponse_Result, error) {
+	pol.resolveDerivedRoles(ctx, loadPolicy)
+	if err := pol.resolveVariables(ctx, loadPolicy); err != nil {
+		return nil, fmt.Errorf("failed to resolve variables: %w", err)
+	}
+
+	return pol.results, nil
+}
+
+func (pol *Policy) resolveDerivedRoles(ctx context.Context, loadPolicy loadPolicyFn) {
 	for policyID, derivedRoles := range pol.derivedRolesToResolve {
 		importedPolicies, ok := pol.derivedRolesImports[policyID]
 		var missingPolicies []string
@@ -188,15 +188,22 @@ func (pol *Policy) Results(ctx context.Context, loadPolicy loadPolicyFn) (map[st
 			}
 		}
 
-		sort.Slice(pol.results[policyID].DerivedRoles, func(x, y int) bool {
-			return pol.results[policyID].DerivedRoles[x].Name < pol.results[policyID].DerivedRoles[y].Name
+		slices.SortFunc(pol.results[policyID].DerivedRoles, func(a, b *responsev1.InspectPoliciesResponse_DerivedRole) int {
+			return cmp.Compare(a.GetName(), b.GetName())
 		})
 	}
+}
 
+func (pol *Policy) resolveVariables(ctx context.Context, loadPolicy loadPolicyFn) error {
 	for policyID, variables := range pol.variablesToResolve {
 		importedPolicies, ok := pol.variableImports[policyID]
+		attrs := make(util.StringSet)
+		for _, attr := range pol.results[policyID].Attributes {
+			attrs[attr.Name] = struct{}{}
+		}
+
 		var missingPolicies []string
-		if ok {
+		if ok { //nolint:nestif
 			for _, importedPolicyID := range importedPolicies {
 				if importedResult, ok := pol.results[importedPolicyID]; ok {
 					for _, importedVariable := range importedResult.Variables {
@@ -209,6 +216,20 @@ func (pol *Policy) Results(ctx context.Context, loadPolicy loadPolicyFn) (map[st
 								Used:   true,
 							})
 							variables[importedVariable.Name] = true
+
+							referencedAttributes := make(map[string]*responsev1.InspectPoliciesResponse_Attribute)
+							if err := pol.referencedAttributesInExpr(importedVariable.Value, referencedAttributes); err != nil {
+								return fmt.Errorf("failed to find referenced attributes in imported variable: %w", err)
+							}
+
+							if len(referencedAttributes) > 0 {
+								for _, attr := range referencedAttributes {
+									if !attrs.Contains(attr.Name) {
+										pol.results[policyID].Attributes = append(pol.results[policyID].Attributes, attr)
+										attrs[attr.Name] = struct{}{}
+									}
+								}
+							}
 						}
 					}
 				} else {
@@ -229,8 +250,21 @@ func (pol *Policy) Results(ctx context.Context, loadPolicy loadPolicyFn) (map[st
 							Source: namer.PolicyKeyFromFQN(wrapper.FQN),
 							Used:   true,
 						})
-
 						variables[importedVarName] = true
+
+						referencedAttributes := make(map[string]*responsev1.InspectPoliciesResponse_Attribute)
+						if err := pol.referencedAttributesInExpr(importedVariable.Value, referencedAttributes); err != nil {
+							return fmt.Errorf("failed to find referenced attributes in imported variable: %w", err)
+						}
+
+						if len(referencedAttributes) > 0 {
+							for _, attr := range referencedAttributes {
+								if !attrs.Contains(attr.Name) {
+									pol.results[policyID].Attributes = append(pol.results[policyID].Attributes, attr)
+									attrs[attr.Name] = struct{}{}
+								}
+							}
+						}
 					}
 				}
 
@@ -252,12 +286,166 @@ func (pol *Policy) Results(ctx context.Context, loadPolicy loadPolicyFn) (map[st
 			}
 		}
 
-		sort.Slice(pol.results[policyID].Variables, func(x, y int) bool {
-			return pol.results[policyID].Variables[x].Name < pol.results[policyID].Variables[y].Name
+		slices.SortFunc(pol.results[policyID].Attributes, func(a, b *responsev1.InspectPoliciesResponse_Attribute) int {
+			return cmp.Compare(a.GetName(), b.GetName())
+		})
+
+		slices.SortFunc(pol.results[policyID].Variables, func(a, b *responsev1.InspectPoliciesResponse_Variable) int {
+			return cmp.Compare(a.GetName(), b.GetName())
 		})
 	}
 
-	return pol.results, nil
+	return nil
+}
+
+// listReferencedAttributes lists the attributes referenced from the conditions and variables in the given policy.
+func (pol *Policy) listReferencedAttributes(p *policyv1.Policy) ([]*responsev1.InspectPoliciesResponse_Attribute, error) {
+	if p == nil {
+		return nil, nil
+	}
+
+	referencedAttributes := make(map[string]*responsev1.InspectPoliciesResponse_Attribute)
+	switch pt := p.PolicyType.(type) {
+	case *policyv1.Policy_DerivedRoles:
+		for _, def := range pt.DerivedRoles.Definitions {
+			if def.Condition == nil {
+				continue
+			}
+
+			if err := pol.referencedAttributesInCondition(def.Condition, referencedAttributes); err != nil {
+				return nil, fmt.Errorf("failed to find referenced attributes in the derived role definition: %w", err)
+			}
+		}
+
+		if pt.DerivedRoles.Variables != nil {
+			for _, expr := range pt.DerivedRoles.Variables.Local {
+				if err := pol.referencedAttributesInExpr(expr, referencedAttributes); err != nil {
+					return nil, fmt.Errorf("failed to find referenced attributes in the derived roles local variable: %w", err)
+				}
+			}
+		}
+	case *policyv1.Policy_ExportVariables:
+		for _, expr := range pt.ExportVariables.Definitions {
+			if err := pol.referencedAttributesInExpr(expr, referencedAttributes); err != nil {
+				return nil, fmt.Errorf("failed to find referenced attributes in exported variable definition: %w", err)
+			}
+		}
+	case *policyv1.Policy_PrincipalPolicy:
+		for _, rule := range pt.PrincipalPolicy.Rules {
+			for _, action := range rule.Actions {
+				if action.Condition == nil {
+					continue
+				}
+
+				if err := pol.referencedAttributesInCondition(action.Condition, referencedAttributes); err != nil {
+					return nil, fmt.Errorf("failed to find referenced attributes in the principal policy rule: %w", err)
+				}
+			}
+		}
+
+		if pt.PrincipalPolicy.Variables != nil {
+			for _, expr := range pt.PrincipalPolicy.Variables.Local {
+				if err := pol.referencedAttributesInExpr(expr, referencedAttributes); err != nil {
+					return nil, fmt.Errorf("failed to find referenced attributes in the principal policy local variable: %w", err)
+				}
+			}
+		}
+	case *policyv1.Policy_ResourcePolicy:
+		for _, rule := range pt.ResourcePolicy.Rules {
+			if rule.Condition == nil {
+				continue
+			}
+
+			if err := pol.referencedAttributesInCondition(rule.Condition, referencedAttributes); err != nil {
+				return nil, fmt.Errorf("failed to find referenced attributes in the resource policy rule: %w", err)
+			}
+		}
+
+		if pt.ResourcePolicy.Variables != nil {
+			for _, expr := range pt.ResourcePolicy.Variables.Local {
+				if err := pol.referencedAttributesInExpr(expr, referencedAttributes); err != nil {
+					return nil, fmt.Errorf("failed to find referenced attributes in the resource policy local variable: %w", err)
+				}
+			}
+		}
+	}
+
+	attributes := make([]*responsev1.InspectPoliciesResponse_Attribute, 0, len(referencedAttributes))
+	for _, attr := range referencedAttributes {
+		attributes = append(attributes, &responsev1.InspectPoliciesResponse_Attribute{
+			Name: attr.Name,
+			Kind: attr.Kind,
+		})
+	}
+
+	if len(attributes) > 0 {
+		slices.SortFunc(attributes, func(a, b *responsev1.InspectPoliciesResponse_Attribute) int {
+			return cmp.Compare(a.GetName(), b.GetName())
+		})
+	}
+
+	return attributes, nil
+}
+
+// listReferencedDerivedRoles lists the referenced derived roles in the given resource policy.
+func (pol *Policy) listReferencedDerivedRoles(rp *policyv1.ResourcePolicy) map[string]struct{} {
+	if rp == nil {
+		return nil
+	}
+
+	derivedRoles := make(map[string]struct{})
+	for _, rule := range rp.Rules {
+		for _, derivedRole := range rule.DerivedRoles {
+			derivedRoles[derivedRole] = struct{}{}
+		}
+	}
+
+	return derivedRoles
+}
+
+// listReferencedVariables lists the variables referenced from the conditions in the given policy.
+func (pol *Policy) listReferencedVariables(p *policyv1.Policy) (map[string]*responsev1.InspectPoliciesResponse_Variable, error) {
+	if p == nil {
+		return nil, nil
+	}
+
+	referencedVariables := make(map[string]*responsev1.InspectPoliciesResponse_Variable)
+	switch pt := p.PolicyType.(type) {
+	case *policyv1.Policy_DerivedRoles:
+		for _, def := range pt.DerivedRoles.Definitions {
+			if def.Condition == nil {
+				continue
+			}
+
+			if err := pol.referencedVariablesInCondition(def.Condition, referencedVariables); err != nil {
+				return nil, fmt.Errorf("failed to find referenced variables in the derived role definition: %w", err)
+			}
+		}
+	case *policyv1.Policy_PrincipalPolicy:
+		for _, rule := range pt.PrincipalPolicy.Rules {
+			for _, action := range rule.Actions {
+				if action.Condition == nil {
+					continue
+				}
+
+				if err := pol.referencedVariablesInCondition(action.Condition, referencedVariables); err != nil {
+					return nil, fmt.Errorf("failed to find referenced variables in the principal policy rule: %w", err)
+				}
+			}
+		}
+	case *policyv1.Policy_ResourcePolicy:
+		for _, rule := range pt.ResourcePolicy.Rules {
+			if rule.Condition == nil {
+				continue
+			}
+
+			if err := pol.referencedVariablesInCondition(rule.Condition, referencedVariables); err != nil {
+				return nil, fmt.Errorf("failed to find referenced variables in the resource policy rule: %w", err)
+			}
+		}
+	}
+
+	return referencedVariables, nil
 }
 
 // listImports lists the derived roles and export variables imported by the given policy.
@@ -296,81 +484,62 @@ func (pol *Policy) listImports(p *policyv1.Policy) (derivedRoleImports, variable
 	return derivedRoleImports, variableImports
 }
 
-// inspectDefinitionsAndRules inspects the definitions and rules in the given policy to find references to the derived roles and variables.
-func (pol *Policy) inspectDefinitionsAndRules(p *policyv1.Policy) (referencedDerivedRoles, referencedVariables map[string]struct{}, err error) {
-	referencedDerivedRoles = make(map[string]struct{})
-	referencedVariables = make(map[string]struct{})
-	switch pt := p.PolicyType.(type) {
-	case *policyv1.Policy_DerivedRoles:
-		for _, def := range pt.DerivedRoles.Definitions {
-			if def.Condition == nil {
-				continue
-			}
-
-			if err := pol.referencedVariableNamesInCondition(def.Condition, referencedVariables); err != nil {
-				return nil, nil, fmt.Errorf("failed to find referenced variable names in condition: %w", err)
-			}
-		}
-	case *policyv1.Policy_PrincipalPolicy:
-		for _, rule := range pt.PrincipalPolicy.Rules {
-			for _, action := range rule.Actions {
-				if action.Condition == nil {
-					continue
-				}
-
-				if err := pol.referencedVariableNamesInCondition(action.Condition, referencedVariables); err != nil {
-					return nil, nil, fmt.Errorf("failed to find referenced variable names in condition: %w", err)
-				}
-			}
-		}
-	case *policyv1.Policy_ResourcePolicy:
-		for _, rule := range pt.ResourcePolicy.Rules {
-			pol.referencedDerivedRolesInResourceRule(rule, referencedDerivedRoles)
-
-			if rule.Condition == nil {
-				continue
-			}
-
-			if err := pol.referencedVariableNamesInCondition(rule.Condition, referencedVariables); err != nil {
-				return nil, nil, fmt.Errorf("failed to find referenced variable names in condition: %w", err)
-			}
-		}
+func (pol *Policy) referencedAttributesInExpr(expr string, referencedAttributes map[string]*responsev1.InspectPoliciesResponse_Attribute) error {
+	c := &policyv1.Condition{
+		Condition: &policyv1.Condition_Match{
+			Match: &policyv1.Match{
+				Op: &policyv1.Match_Expr{
+					Expr: expr,
+				},
+			},
+		},
 	}
 
-	return referencedDerivedRoles, referencedVariables, nil
-}
-
-func (pol *Policy) referencedDerivedRolesInResourceRule(rule *policyv1.ResourceRule, out map[string]struct{}) {
-	for _, derivedRole := range rule.DerivedRoles {
-		out[derivedRole] = struct{}{}
-	}
-}
-
-func (pol *Policy) referencedVariableNamesInCondition(condition *policyv1.Condition, out map[string]struct{}) error {
-	c, err := compile.Condition(condition)
-	if err != nil {
-		return fmt.Errorf("failed to compile condition: %w", err)
-	}
-
-	if err := pol.referencedVariableNamesInCompiledCondition(c, out); err != nil {
-		return fmt.Errorf("failed to find referenced variable names in compiled condition: %w", err)
+	if err := pol.referencedAttributesInCondition(c, referencedAttributes); err != nil {
+		return fmt.Errorf("failed to find referenced attributes in the expression: %w", err)
 	}
 
 	return nil
 }
 
-func (pol *Policy) referencedVariableNamesInCompiledCondition(condition *runtimev1.Condition, out map[string]struct{}) error {
+func (pol *Policy) referencedAttributesInCondition(condition *policyv1.Condition, referencedAttributes map[string]*responsev1.InspectPoliciesResponse_Attribute) error {
+	c, err := compile.Condition(condition)
+	if err != nil {
+		return fmt.Errorf("failed to compile the condition: %w", err)
+	}
+
+	if err := referencedAttributesInCompiledCondition(c, referencedAttributes); err != nil {
+		return fmt.Errorf("failed to find referenced attributes in the compiled condition: %w", err)
+	}
+
+	return nil
+}
+
+func (pol *Policy) referencedVariablesInCondition(condition *policyv1.Condition, referencedVariables map[string]*responsev1.InspectPoliciesResponse_Variable) error {
+	c, err := compile.Condition(condition)
+	if err != nil {
+		return fmt.Errorf("failed to compile the condition: %w", err)
+	}
+
+	if err := pol.referencedVariablesInCompiledCondition(c, referencedVariables); err != nil {
+		return fmt.Errorf("failed to find referenced variables in the compiled condition: %w", err)
+	}
+
+	return nil
+}
+
+func (pol *Policy) referencedVariablesInCompiledCondition(condition *runtimev1.Condition, out map[string]*responsev1.InspectPoliciesResponse_Variable) error {
 	switch op := condition.Op.(type) {
 	case *runtimev1.Condition_All:
 		for _, condition := range op.All.Expr {
-			if err := pol.referencedVariableNamesInCompiledCondition(condition, out); err != nil {
-				return fmt.Errorf("failed to find referenced variable names in all condition: %w", err)
+			if err := pol.referencedVariablesInCompiledCondition(condition, out); err != nil {
+				return fmt.Errorf("failed to find referenced variables in the 'all' expression: %w", err)
 			}
 		}
 	case *runtimev1.Condition_Any:
 		for _, condition := range op.Any.Expr {
-			if err := pol.referencedVariableNamesInCompiledCondition(condition, out); err != nil {
-				return fmt.Errorf("failed to find referenced variable names in any condition: %w", err)
+			if err := pol.referencedVariablesInCompiledCondition(condition, out); err != nil {
+				return fmt.Errorf("failed to find referenced variables in the 'any' expression: %w", err)
 			}
 		}
 	case *runtimev1.Condition_Expr:
@@ -379,14 +548,23 @@ func (pol *Policy) referencedVariableNamesInCompiledCondition(condition *runtime
 			return fmt.Errorf("failed to convert checked expression %s to AST: %w", op.Expr.Checked, err)
 		}
 
-		action := func(varName string) {
-			out[varName] = struct{}{}
-		}
-		ast.PreOrderVisit(exprAST.Expr(), compile.VariableVisitor(action))
+		ast.PreOrderVisit(
+			exprAST.Expr(),
+			variableVisitor(
+				func(name, value string) {
+					out[name] = &responsev1.InspectPoliciesResponse_Variable{
+						Name:  name,
+						Value: value,
+						Kind:  responsev1.InspectPoliciesResponse_Variable_KIND_UNKNOWN,
+						Used:  true,
+					}
+				},
+			),
+		)
 	case *runtimev1.Condition_None:
 		for _, condition := range op.None.Expr {
-			if err := pol.referencedVariableNamesInCompiledCondition(condition, out); err != nil {
-				return fmt.Errorf("failed to find referenced variable names in none condition: %w", err)
+			if err := pol.referencedVariablesInCompiledCondition(condition, out); err != nil {
+				return fmt.Errorf("failed to find referenced variables in the 'none' expression: %w", err)
 			}
 		}
 	}
