@@ -29,6 +29,61 @@ const (
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
 
+type Kind int32
+
+const (
+	Kind_KIND_UNSPECIFIED      Kind = 0
+	Kind_KIND_DERIVED_ROLES    Kind = 1
+	Kind_KIND_EXPORT_VARIABLES Kind = 2
+	Kind_KIND_PRINCIPAL        Kind = 3
+	Kind_KIND_RESOURCE         Kind = 4
+)
+
+// Enum value maps for Kind.
+var (
+	Kind_name = map[int32]string{
+		0: "KIND_UNSPECIFIED",
+		1: "KIND_DERIVED_ROLES",
+		2: "KIND_EXPORT_VARIABLES",
+		3: "KIND_PRINCIPAL",
+		4: "KIND_RESOURCE",
+	}
+	Kind_value = map[string]int32{
+		"KIND_UNSPECIFIED":      0,
+		"KIND_DERIVED_ROLES":    1,
+		"KIND_EXPORT_VARIABLES": 2,
+		"KIND_PRINCIPAL":        3,
+		"KIND_RESOURCE":         4,
+	}
+)
+
+func (x Kind) Enum() *Kind {
+	p := new(Kind)
+	*p = x
+	return p
+}
+
+func (x Kind) String() string {
+	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
+}
+
+func (Kind) Descriptor() protoreflect.EnumDescriptor {
+	return file_cerbos_policy_v1_policy_proto_enumTypes[0].Descriptor()
+}
+
+func (Kind) Type() protoreflect.EnumType {
+	return &file_cerbos_policy_v1_policy_proto_enumTypes[0]
+}
+
+func (x Kind) Number() protoreflect.EnumNumber {
+	return protoreflect.EnumNumber(x)
+}
+
+// Deprecated: Use Kind.Descriptor instead.
+func (Kind) EnumDescriptor() ([]byte, []int) {
+	return file_cerbos_policy_v1_policy_proto_rawDescGZIP(), []int{0}
+}
+
 type TestResults_Result int32
 
 const (
@@ -68,11 +123,11 @@ func (x TestResults_Result) String() string {
 }
 
 func (TestResults_Result) Descriptor() protoreflect.EnumDescriptor {
-	return file_cerbos_policy_v1_policy_proto_enumTypes[0].Descriptor()
+	return file_cerbos_policy_v1_policy_proto_enumTypes[1].Descriptor()
 }
 
 func (TestResults_Result) Type() protoreflect.EnumType {
-	return &file_cerbos_policy_v1_policy_proto_enumTypes[0]
+	return &file_cerbos_policy_v1_policy_proto_enumTypes[1]
 }
 
 func (x TestResults_Result) Number() protoreflect.EnumNumber {
@@ -4176,15 +4231,22 @@ var file_cerbos_policy_v1_policy_proto_rawDesc = []byte{
 	0x45, 0x44, 0x10, 0x01, 0x12, 0x11, 0x0a, 0x0d, 0x52, 0x45, 0x53, 0x55, 0x4c, 0x54, 0x5f, 0x50,
 	0x41, 0x53, 0x53, 0x45, 0x44, 0x10, 0x02, 0x12, 0x11, 0x0a, 0x0d, 0x52, 0x45, 0x53, 0x55, 0x4c,
 	0x54, 0x5f, 0x46, 0x41, 0x49, 0x4c, 0x45, 0x44, 0x10, 0x03, 0x12, 0x12, 0x0a, 0x0e, 0x52, 0x45,
-	0x53, 0x55, 0x4c, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x4f, 0x52, 0x45, 0x44, 0x10, 0x04, 0x42, 0x6f,
-	0x0a, 0x18, 0x64, 0x65, 0x76, 0x2e, 0x63, 0x65, 0x72, 0x62, 0x6f, 0x73, 0x2e, 0x61, 0x70, 0x69,
-	0x2e, 0x76, 0x31, 0x2e, 0x70, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x5a, 0x3c, 0x67, 0x69, 0x74, 0x68,
-	0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x63, 0x65, 0x72, 0x62, 0x6f, 0x73, 0x2f, 0x63, 0x65,
-	0x72, 0x62, 0x6f, 0x73, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x67, 0x65, 0x6e, 0x70, 0x62, 0x2f, 0x63,
-	0x65, 0x72, 0x62, 0x6f, 0x73, 0x2f, 0x70, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x2f, 0x76, 0x31, 0x3b,
-	0x70, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x76, 0x31, 0xaa, 0x02, 0x14, 0x43, 0x65, 0x72, 0x62, 0x6f,
-	0x73, 0x2e, 0x41, 0x70, 0x69, 0x2e, 0x56, 0x31, 0x2e, 0x50, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x62,
-	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x53, 0x55, 0x4c, 0x54, 0x5f, 0x45, 0x52, 0x52, 0x4f, 0x52, 0x45, 0x44, 0x10, 0x04, 0x2a, 0x76,
+	0x0a, 0x04, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x14, 0x0a, 0x10, 0x4b, 0x49, 0x4e, 0x44, 0x5f, 0x55,
+	0x4e, 0x53, 0x50, 0x45, 0x43, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x00, 0x12, 0x16, 0x0a, 0x12,
+	0x4b, 0x49, 0x4e, 0x44, 0x5f, 0x44, 0x45, 0x52, 0x49, 0x56, 0x45, 0x44, 0x5f, 0x52, 0x4f, 0x4c,
+	0x45, 0x53, 0x10, 0x01, 0x12, 0x19, 0x0a, 0x15, 0x4b, 0x49, 0x4e, 0x44, 0x5f, 0x45, 0x58, 0x50,
+	0x4f, 0x52, 0x54, 0x5f, 0x56, 0x41, 0x52, 0x49, 0x41, 0x42, 0x4c, 0x45, 0x53, 0x10, 0x02, 0x12,
+	0x12, 0x0a, 0x0e, 0x4b, 0x49, 0x4e, 0x44, 0x5f, 0x50, 0x52, 0x49, 0x4e, 0x43, 0x49, 0x50, 0x41,
+	0x4c, 0x10, 0x03, 0x12, 0x11, 0x0a, 0x0d, 0x4b, 0x49, 0x4e, 0x44, 0x5f, 0x52, 0x45, 0x53, 0x4f,
+	0x55, 0x52, 0x43, 0x45, 0x10, 0x04, 0x42, 0x6f, 0x0a, 0x18, 0x64, 0x65, 0x76, 0x2e, 0x63, 0x65,
+	0x72, 0x62, 0x6f, 0x73, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x70, 0x6f, 0x6c, 0x69,
+	0x63, 0x79, 0x5a, 0x3c, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x63,
+	0x65, 0x72, 0x62, 0x6f, 0x73, 0x2f, 0x63, 0x65, 0x72, 0x62, 0x6f, 0x73, 0x2f, 0x61, 0x70, 0x69,
+	0x2f, 0x67, 0x65, 0x6e, 0x70, 0x62, 0x2f, 0x63, 0x65, 0x72, 0x62, 0x6f, 0x73, 0x2f, 0x70, 0x6f,
+	0x6c, 0x69, 0x63, 0x79, 0x2f, 0x76, 0x31, 0x3b, 0x70, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x76, 0x31,
+	0xaa, 0x02, 0x14, 0x43, 0x65, 0x72, 0x62, 0x6f, 0x73, 0x2e, 0x41, 0x70, 0x69, 0x2e, 0x56, 0x31,
+	0x2e, 0x50, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -4199,185 +4261,186 @@ func file_cerbos_policy_v1_policy_proto_rawDescGZIP() []byte {
 	return file_cerbos_policy_v1_policy_proto_rawDescData
 }
 
-var file_cerbos_policy_v1_policy_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
+var file_cerbos_policy_v1_policy_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
 var file_cerbos_policy_v1_policy_proto_msgTypes = make([]protoimpl.MessageInfo, 65)
 var file_cerbos_policy_v1_policy_proto_goTypes = []any{
-	(TestResults_Result)(0),              // 0: cerbos.policy.v1.TestResults.Result
-	(*Policy)(nil),                       // 1: cerbos.policy.v1.Policy
-	(*SourceAttributes)(nil),             // 2: cerbos.policy.v1.SourceAttributes
-	(*Metadata)(nil),                     // 3: cerbos.policy.v1.Metadata
-	(*ResourcePolicy)(nil),               // 4: cerbos.policy.v1.ResourcePolicy
-	(*ResourceRule)(nil),                 // 5: cerbos.policy.v1.ResourceRule
-	(*RolePolicy)(nil),                   // 6: cerbos.policy.v1.RolePolicy
-	(*RoleRule)(nil),                     // 7: cerbos.policy.v1.RoleRule
-	(*PrincipalPolicy)(nil),              // 8: cerbos.policy.v1.PrincipalPolicy
-	(*PrincipalRule)(nil),                // 9: cerbos.policy.v1.PrincipalRule
-	(*DerivedRoles)(nil),                 // 10: cerbos.policy.v1.DerivedRoles
-	(*RoleDef)(nil),                      // 11: cerbos.policy.v1.RoleDef
-	(*ExportVariables)(nil),              // 12: cerbos.policy.v1.ExportVariables
-	(*Variables)(nil),                    // 13: cerbos.policy.v1.Variables
-	(*Condition)(nil),                    // 14: cerbos.policy.v1.Condition
-	(*Match)(nil),                        // 15: cerbos.policy.v1.Match
-	(*Output)(nil),                       // 16: cerbos.policy.v1.Output
-	(*Schemas)(nil),                      // 17: cerbos.policy.v1.Schemas
-	(*TestFixture)(nil),                  // 18: cerbos.policy.v1.TestFixture
-	(*TestOptions)(nil),                  // 19: cerbos.policy.v1.TestOptions
-	(*TestSuite)(nil),                    // 20: cerbos.policy.v1.TestSuite
-	(*TestTable)(nil),                    // 21: cerbos.policy.v1.TestTable
-	(*Test)(nil),                         // 22: cerbos.policy.v1.Test
-	(*TestResults)(nil),                  // 23: cerbos.policy.v1.TestResults
-	nil,                                  // 24: cerbos.policy.v1.Policy.VariablesEntry
-	nil,                                  // 25: cerbos.policy.v1.SourceAttributes.AttributesEntry
-	nil,                                  // 26: cerbos.policy.v1.Metadata.AnnotationsEntry
-	(*PrincipalRule_Action)(nil),         // 27: cerbos.policy.v1.PrincipalRule.Action
-	nil,                                  // 28: cerbos.policy.v1.ExportVariables.DefinitionsEntry
-	nil,                                  // 29: cerbos.policy.v1.Variables.LocalEntry
-	(*Match_ExprList)(nil),               // 30: cerbos.policy.v1.Match.ExprList
-	(*Output_When)(nil),                  // 31: cerbos.policy.v1.Output.When
-	(*Schemas_IgnoreWhen)(nil),           // 32: cerbos.policy.v1.Schemas.IgnoreWhen
-	(*Schemas_Schema)(nil),               // 33: cerbos.policy.v1.Schemas.Schema
-	(*TestFixture_Principals)(nil),       // 34: cerbos.policy.v1.TestFixture.Principals
-	(*TestFixture_Resources)(nil),        // 35: cerbos.policy.v1.TestFixture.Resources
-	(*TestFixture_AuxData)(nil),          // 36: cerbos.policy.v1.TestFixture.AuxData
-	nil,                                  // 37: cerbos.policy.v1.TestFixture.Principals.PrincipalsEntry
-	nil,                                  // 38: cerbos.policy.v1.TestFixture.Resources.ResourcesEntry
-	nil,                                  // 39: cerbos.policy.v1.TestFixture.AuxData.AuxDataEntry
-	nil,                                  // 40: cerbos.policy.v1.TestOptions.GlobalsEntry
-	nil,                                  // 41: cerbos.policy.v1.TestSuite.PrincipalsEntry
-	nil,                                  // 42: cerbos.policy.v1.TestSuite.ResourcesEntry
-	nil,                                  // 43: cerbos.policy.v1.TestSuite.AuxDataEntry
-	(*TestTable_Input)(nil),              // 44: cerbos.policy.v1.TestTable.Input
-	(*TestTable_OutputExpectations)(nil), // 45: cerbos.policy.v1.TestTable.OutputExpectations
-	(*TestTable_Expectation)(nil),        // 46: cerbos.policy.v1.TestTable.Expectation
-	nil,                                  // 47: cerbos.policy.v1.TestTable.Expectation.ActionsEntry
-	(*Test_TestName)(nil),                // 48: cerbos.policy.v1.Test.TestName
-	(*Test_OutputEntries)(nil),           // 49: cerbos.policy.v1.Test.OutputEntries
-	nil,                                  // 50: cerbos.policy.v1.Test.ExpectedEntry
-	nil,                                  // 51: cerbos.policy.v1.Test.ExpectedOutputsEntry
-	nil,                                  // 52: cerbos.policy.v1.Test.OutputEntries.EntriesEntry
-	(*TestResults_Tally)(nil),            // 53: cerbos.policy.v1.TestResults.Tally
-	(*TestResults_Summary)(nil),          // 54: cerbos.policy.v1.TestResults.Summary
-	(*TestResults_Suite)(nil),            // 55: cerbos.policy.v1.TestResults.Suite
-	(*TestResults_TestCase)(nil),         // 56: cerbos.policy.v1.TestResults.TestCase
-	(*TestResults_Principal)(nil),        // 57: cerbos.policy.v1.TestResults.Principal
-	(*TestResults_Resource)(nil),         // 58: cerbos.policy.v1.TestResults.Resource
-	(*TestResults_Action)(nil),           // 59: cerbos.policy.v1.TestResults.Action
-	(*TestResults_Details)(nil),          // 60: cerbos.policy.v1.TestResults.Details
-	(*TestResults_OutputFailure)(nil),    // 61: cerbos.policy.v1.TestResults.OutputFailure
-	(*TestResults_Failure)(nil),          // 62: cerbos.policy.v1.TestResults.Failure
-	(*TestResults_Success)(nil),          // 63: cerbos.policy.v1.TestResults.Success
-	(*TestResults_OutputFailure_MismatchedValue)(nil), // 64: cerbos.policy.v1.TestResults.OutputFailure.MismatchedValue
-	(*TestResults_OutputFailure_MissingValue)(nil),    // 65: cerbos.policy.v1.TestResults.OutputFailure.MissingValue
-	(*wrapperspb.UInt64Value)(nil),                    // 66: google.protobuf.UInt64Value
-	(v1.Effect)(0),                                    // 67: cerbos.effect.v1.Effect
-	(*timestamppb.Timestamp)(nil),                     // 68: google.protobuf.Timestamp
-	(*v11.CheckInput)(nil),                            // 69: cerbos.engine.v1.CheckInput
-	(*structpb.Value)(nil),                            // 70: google.protobuf.Value
-	(*v11.Principal)(nil),                             // 71: cerbos.engine.v1.Principal
-	(*v11.Resource)(nil),                              // 72: cerbos.engine.v1.Resource
-	(*v11.AuxData)(nil),                               // 73: cerbos.engine.v1.AuxData
-	(*v11.OutputEntry)(nil),                           // 74: cerbos.engine.v1.OutputEntry
-	(*v11.Trace)(nil),                                 // 75: cerbos.engine.v1.Trace
+	(Kind)(0),                            // 0: cerbos.policy.v1.Kind
+	(TestResults_Result)(0),              // 1: cerbos.policy.v1.TestResults.Result
+	(*Policy)(nil),                       // 2: cerbos.policy.v1.Policy
+	(*SourceAttributes)(nil),             // 3: cerbos.policy.v1.SourceAttributes
+	(*Metadata)(nil),                     // 4: cerbos.policy.v1.Metadata
+	(*ResourcePolicy)(nil),               // 5: cerbos.policy.v1.ResourcePolicy
+	(*ResourceRule)(nil),                 // 6: cerbos.policy.v1.ResourceRule
+	(*RolePolicy)(nil),                   // 7: cerbos.policy.v1.RolePolicy
+	(*RoleRule)(nil),                     // 8: cerbos.policy.v1.RoleRule
+	(*PrincipalPolicy)(nil),              // 9: cerbos.policy.v1.PrincipalPolicy
+	(*PrincipalRule)(nil),                // 10: cerbos.policy.v1.PrincipalRule
+	(*DerivedRoles)(nil),                 // 11: cerbos.policy.v1.DerivedRoles
+	(*RoleDef)(nil),                      // 12: cerbos.policy.v1.RoleDef
+	(*ExportVariables)(nil),              // 13: cerbos.policy.v1.ExportVariables
+	(*Variables)(nil),                    // 14: cerbos.policy.v1.Variables
+	(*Condition)(nil),                    // 15: cerbos.policy.v1.Condition
+	(*Match)(nil),                        // 16: cerbos.policy.v1.Match
+	(*Output)(nil),                       // 17: cerbos.policy.v1.Output
+	(*Schemas)(nil),                      // 18: cerbos.policy.v1.Schemas
+	(*TestFixture)(nil),                  // 19: cerbos.policy.v1.TestFixture
+	(*TestOptions)(nil),                  // 20: cerbos.policy.v1.TestOptions
+	(*TestSuite)(nil),                    // 21: cerbos.policy.v1.TestSuite
+	(*TestTable)(nil),                    // 22: cerbos.policy.v1.TestTable
+	(*Test)(nil),                         // 23: cerbos.policy.v1.Test
+	(*TestResults)(nil),                  // 24: cerbos.policy.v1.TestResults
+	nil,                                  // 25: cerbos.policy.v1.Policy.VariablesEntry
+	nil,                                  // 26: cerbos.policy.v1.SourceAttributes.AttributesEntry
+	nil,                                  // 27: cerbos.policy.v1.Metadata.AnnotationsEntry
+	(*PrincipalRule_Action)(nil),         // 28: cerbos.policy.v1.PrincipalRule.Action
+	nil,                                  // 29: cerbos.policy.v1.ExportVariables.DefinitionsEntry
+	nil,                                  // 30: cerbos.policy.v1.Variables.LocalEntry
+	(*Match_ExprList)(nil),               // 31: cerbos.policy.v1.Match.ExprList
+	(*Output_When)(nil),                  // 32: cerbos.policy.v1.Output.When
+	(*Schemas_IgnoreWhen)(nil),           // 33: cerbos.policy.v1.Schemas.IgnoreWhen
+	(*Schemas_Schema)(nil),               // 34: cerbos.policy.v1.Schemas.Schema
+	(*TestFixture_Principals)(nil),       // 35: cerbos.policy.v1.TestFixture.Principals
+	(*TestFixture_Resources)(nil),        // 36: cerbos.policy.v1.TestFixture.Resources
+	(*TestFixture_AuxData)(nil),          // 37: cerbos.policy.v1.TestFixture.AuxData
+	nil,                                  // 38: cerbos.policy.v1.TestFixture.Principals.PrincipalsEntry
+	nil,                                  // 39: cerbos.policy.v1.TestFixture.Resources.ResourcesEntry
+	nil,                                  // 40: cerbos.policy.v1.TestFixture.AuxData.AuxDataEntry
+	nil,                                  // 41: cerbos.policy.v1.TestOptions.GlobalsEntry
+	nil,                                  // 42: cerbos.policy.v1.TestSuite.PrincipalsEntry
+	nil,                                  // 43: cerbos.policy.v1.TestSuite.ResourcesEntry
+	nil,                                  // 44: cerbos.policy.v1.TestSuite.AuxDataEntry
+	(*TestTable_Input)(nil),              // 45: cerbos.policy.v1.TestTable.Input
+	(*TestTable_OutputExpectations)(nil), // 46: cerbos.policy.v1.TestTable.OutputExpectations
+	(*TestTable_Expectation)(nil),        // 47: cerbos.policy.v1.TestTable.Expectation
+	nil,                                  // 48: cerbos.policy.v1.TestTable.Expectation.ActionsEntry
+	(*Test_TestName)(nil),                // 49: cerbos.policy.v1.Test.TestName
+	(*Test_OutputEntries)(nil),           // 50: cerbos.policy.v1.Test.OutputEntries
+	nil,                                  // 51: cerbos.policy.v1.Test.ExpectedEntry
+	nil,                                  // 52: cerbos.policy.v1.Test.ExpectedOutputsEntry
+	nil,                                  // 53: cerbos.policy.v1.Test.OutputEntries.EntriesEntry
+	(*TestResults_Tally)(nil),            // 54: cerbos.policy.v1.TestResults.Tally
+	(*TestResults_Summary)(nil),          // 55: cerbos.policy.v1.TestResults.Summary
+	(*TestResults_Suite)(nil),            // 56: cerbos.policy.v1.TestResults.Suite
+	(*TestResults_TestCase)(nil),         // 57: cerbos.policy.v1.TestResults.TestCase
+	(*TestResults_Principal)(nil),        // 58: cerbos.policy.v1.TestResults.Principal
+	(*TestResults_Resource)(nil),         // 59: cerbos.policy.v1.TestResults.Resource
+	(*TestResults_Action)(nil),           // 60: cerbos.policy.v1.TestResults.Action
+	(*TestResults_Details)(nil),          // 61: cerbos.policy.v1.TestResults.Details
+	(*TestResults_OutputFailure)(nil),    // 62: cerbos.policy.v1.TestResults.OutputFailure
+	(*TestResults_Failure)(nil),          // 63: cerbos.policy.v1.TestResults.Failure
+	(*TestResults_Success)(nil),          // 64: cerbos.policy.v1.TestResults.Success
+	(*TestResults_OutputFailure_MismatchedValue)(nil), // 65: cerbos.policy.v1.TestResults.OutputFailure.MismatchedValue
+	(*TestResults_OutputFailure_MissingValue)(nil),    // 66: cerbos.policy.v1.TestResults.OutputFailure.MissingValue
+	(*wrapperspb.UInt64Value)(nil),                    // 67: google.protobuf.UInt64Value
+	(v1.Effect)(0),                                    // 68: cerbos.effect.v1.Effect
+	(*timestamppb.Timestamp)(nil),                     // 69: google.protobuf.Timestamp
+	(*v11.CheckInput)(nil),                            // 70: cerbos.engine.v1.CheckInput
+	(*structpb.Value)(nil),                            // 71: google.protobuf.Value
+	(*v11.Principal)(nil),                             // 72: cerbos.engine.v1.Principal
+	(*v11.Resource)(nil),                              // 73: cerbos.engine.v1.Resource
+	(*v11.AuxData)(nil),                               // 74: cerbos.engine.v1.AuxData
+	(*v11.OutputEntry)(nil),                           // 75: cerbos.engine.v1.OutputEntry
+	(*v11.Trace)(nil),                                 // 76: cerbos.engine.v1.Trace
 }
 var file_cerbos_policy_v1_policy_proto_depIdxs = []int32{
-	3,  // 0: cerbos.policy.v1.Policy.metadata:type_name -> cerbos.policy.v1.Metadata
-	4,  // 1: cerbos.policy.v1.Policy.resource_policy:type_name -> cerbos.policy.v1.ResourcePolicy
-	8,  // 2: cerbos.policy.v1.Policy.principal_policy:type_name -> cerbos.policy.v1.PrincipalPolicy
-	10, // 3: cerbos.policy.v1.Policy.derived_roles:type_name -> cerbos.policy.v1.DerivedRoles
-	12, // 4: cerbos.policy.v1.Policy.export_variables:type_name -> cerbos.policy.v1.ExportVariables
-	6,  // 5: cerbos.policy.v1.Policy.role_policy:type_name -> cerbos.policy.v1.RolePolicy
-	24, // 6: cerbos.policy.v1.Policy.variables:type_name -> cerbos.policy.v1.Policy.VariablesEntry
-	25, // 7: cerbos.policy.v1.SourceAttributes.attributes:type_name -> cerbos.policy.v1.SourceAttributes.AttributesEntry
-	26, // 8: cerbos.policy.v1.Metadata.annotations:type_name -> cerbos.policy.v1.Metadata.AnnotationsEntry
-	66, // 9: cerbos.policy.v1.Metadata.hash:type_name -> google.protobuf.UInt64Value
-	2,  // 10: cerbos.policy.v1.Metadata.source_attributes:type_name -> cerbos.policy.v1.SourceAttributes
-	5,  // 11: cerbos.policy.v1.ResourcePolicy.rules:type_name -> cerbos.policy.v1.ResourceRule
-	17, // 12: cerbos.policy.v1.ResourcePolicy.schemas:type_name -> cerbos.policy.v1.Schemas
-	13, // 13: cerbos.policy.v1.ResourcePolicy.variables:type_name -> cerbos.policy.v1.Variables
-	14, // 14: cerbos.policy.v1.ResourceRule.condition:type_name -> cerbos.policy.v1.Condition
-	67, // 15: cerbos.policy.v1.ResourceRule.effect:type_name -> cerbos.effect.v1.Effect
-	16, // 16: cerbos.policy.v1.ResourceRule.output:type_name -> cerbos.policy.v1.Output
-	7,  // 17: cerbos.policy.v1.RolePolicy.rules:type_name -> cerbos.policy.v1.RoleRule
-	9,  // 18: cerbos.policy.v1.PrincipalPolicy.rules:type_name -> cerbos.policy.v1.PrincipalRule
-	13, // 19: cerbos.policy.v1.PrincipalPolicy.variables:type_name -> cerbos.policy.v1.Variables
-	27, // 20: cerbos.policy.v1.PrincipalRule.actions:type_name -> cerbos.policy.v1.PrincipalRule.Action
-	11, // 21: cerbos.policy.v1.DerivedRoles.definitions:type_name -> cerbos.policy.v1.RoleDef
-	13, // 22: cerbos.policy.v1.DerivedRoles.variables:type_name -> cerbos.policy.v1.Variables
-	14, // 23: cerbos.policy.v1.RoleDef.condition:type_name -> cerbos.policy.v1.Condition
-	28, // 24: cerbos.policy.v1.ExportVariables.definitions:type_name -> cerbos.policy.v1.ExportVariables.DefinitionsEntry
-	29, // 25: cerbos.policy.v1.Variables.local:type_name -> cerbos.policy.v1.Variables.LocalEntry
-	15, // 26: cerbos.policy.v1.Condition.match:type_name -> cerbos.policy.v1.Match
-	30, // 27: cerbos.policy.v1.Match.all:type_name -> cerbos.policy.v1.Match.ExprList
-	30, // 28: cerbos.policy.v1.Match.any:type_name -> cerbos.policy.v1.Match.ExprList
-	30, // 29: cerbos.policy.v1.Match.none:type_name -> cerbos.policy.v1.Match.ExprList
-	31, // 30: cerbos.policy.v1.Output.when:type_name -> cerbos.policy.v1.Output.When
-	33, // 31: cerbos.policy.v1.Schemas.principal_schema:type_name -> cerbos.policy.v1.Schemas.Schema
-	33, // 32: cerbos.policy.v1.Schemas.resource_schema:type_name -> cerbos.policy.v1.Schemas.Schema
-	68, // 33: cerbos.policy.v1.TestOptions.now:type_name -> google.protobuf.Timestamp
-	40, // 34: cerbos.policy.v1.TestOptions.globals:type_name -> cerbos.policy.v1.TestOptions.GlobalsEntry
-	21, // 35: cerbos.policy.v1.TestSuite.tests:type_name -> cerbos.policy.v1.TestTable
-	41, // 36: cerbos.policy.v1.TestSuite.principals:type_name -> cerbos.policy.v1.TestSuite.PrincipalsEntry
-	42, // 37: cerbos.policy.v1.TestSuite.resources:type_name -> cerbos.policy.v1.TestSuite.ResourcesEntry
-	43, // 38: cerbos.policy.v1.TestSuite.aux_data:type_name -> cerbos.policy.v1.TestSuite.AuxDataEntry
-	19, // 39: cerbos.policy.v1.TestSuite.options:type_name -> cerbos.policy.v1.TestOptions
-	44, // 40: cerbos.policy.v1.TestTable.input:type_name -> cerbos.policy.v1.TestTable.Input
-	46, // 41: cerbos.policy.v1.TestTable.expected:type_name -> cerbos.policy.v1.TestTable.Expectation
-	19, // 42: cerbos.policy.v1.TestTable.options:type_name -> cerbos.policy.v1.TestOptions
-	48, // 43: cerbos.policy.v1.Test.name:type_name -> cerbos.policy.v1.Test.TestName
-	69, // 44: cerbos.policy.v1.Test.input:type_name -> cerbos.engine.v1.CheckInput
-	50, // 45: cerbos.policy.v1.Test.expected:type_name -> cerbos.policy.v1.Test.ExpectedEntry
-	19, // 46: cerbos.policy.v1.Test.options:type_name -> cerbos.policy.v1.TestOptions
-	51, // 47: cerbos.policy.v1.Test.expected_outputs:type_name -> cerbos.policy.v1.Test.ExpectedOutputsEntry
-	55, // 48: cerbos.policy.v1.TestResults.suites:type_name -> cerbos.policy.v1.TestResults.Suite
-	54, // 49: cerbos.policy.v1.TestResults.summary:type_name -> cerbos.policy.v1.TestResults.Summary
-	70, // 50: cerbos.policy.v1.SourceAttributes.AttributesEntry.value:type_name -> google.protobuf.Value
-	14, // 51: cerbos.policy.v1.PrincipalRule.Action.condition:type_name -> cerbos.policy.v1.Condition
-	67, // 52: cerbos.policy.v1.PrincipalRule.Action.effect:type_name -> cerbos.effect.v1.Effect
-	16, // 53: cerbos.policy.v1.PrincipalRule.Action.output:type_name -> cerbos.policy.v1.Output
-	15, // 54: cerbos.policy.v1.Match.ExprList.of:type_name -> cerbos.policy.v1.Match
-	32, // 55: cerbos.policy.v1.Schemas.Schema.ignore_when:type_name -> cerbos.policy.v1.Schemas.IgnoreWhen
-	37, // 56: cerbos.policy.v1.TestFixture.Principals.principals:type_name -> cerbos.policy.v1.TestFixture.Principals.PrincipalsEntry
-	38, // 57: cerbos.policy.v1.TestFixture.Resources.resources:type_name -> cerbos.policy.v1.TestFixture.Resources.ResourcesEntry
-	39, // 58: cerbos.policy.v1.TestFixture.AuxData.aux_data:type_name -> cerbos.policy.v1.TestFixture.AuxData.AuxDataEntry
-	71, // 59: cerbos.policy.v1.TestFixture.Principals.PrincipalsEntry.value:type_name -> cerbos.engine.v1.Principal
-	72, // 60: cerbos.policy.v1.TestFixture.Resources.ResourcesEntry.value:type_name -> cerbos.engine.v1.Resource
-	73, // 61: cerbos.policy.v1.TestFixture.AuxData.AuxDataEntry.value:type_name -> cerbos.engine.v1.AuxData
-	70, // 62: cerbos.policy.v1.TestOptions.GlobalsEntry.value:type_name -> google.protobuf.Value
-	71, // 63: cerbos.policy.v1.TestSuite.PrincipalsEntry.value:type_name -> cerbos.engine.v1.Principal
-	72, // 64: cerbos.policy.v1.TestSuite.ResourcesEntry.value:type_name -> cerbos.engine.v1.Resource
-	73, // 65: cerbos.policy.v1.TestSuite.AuxDataEntry.value:type_name -> cerbos.engine.v1.AuxData
-	74, // 66: cerbos.policy.v1.TestTable.OutputExpectations.expected:type_name -> cerbos.engine.v1.OutputEntry
-	47, // 67: cerbos.policy.v1.TestTable.Expectation.actions:type_name -> cerbos.policy.v1.TestTable.Expectation.ActionsEntry
-	45, // 68: cerbos.policy.v1.TestTable.Expectation.outputs:type_name -> cerbos.policy.v1.TestTable.OutputExpectations
-	67, // 69: cerbos.policy.v1.TestTable.Expectation.ActionsEntry.value:type_name -> cerbos.effect.v1.Effect
-	52, // 70: cerbos.policy.v1.Test.OutputEntries.entries:type_name -> cerbos.policy.v1.Test.OutputEntries.EntriesEntry
-	67, // 71: cerbos.policy.v1.Test.ExpectedEntry.value:type_name -> cerbos.effect.v1.Effect
-	49, // 72: cerbos.policy.v1.Test.ExpectedOutputsEntry.value:type_name -> cerbos.policy.v1.Test.OutputEntries
-	70, // 73: cerbos.policy.v1.Test.OutputEntries.EntriesEntry.value:type_name -> google.protobuf.Value
-	0,  // 74: cerbos.policy.v1.TestResults.Tally.result:type_name -> cerbos.policy.v1.TestResults.Result
-	0,  // 75: cerbos.policy.v1.TestResults.Summary.overall_result:type_name -> cerbos.policy.v1.TestResults.Result
-	53, // 76: cerbos.policy.v1.TestResults.Summary.result_counts:type_name -> cerbos.policy.v1.TestResults.Tally
-	57, // 77: cerbos.policy.v1.TestResults.Suite.principals:type_name -> cerbos.policy.v1.TestResults.Principal
-	54, // 78: cerbos.policy.v1.TestResults.Suite.summary:type_name -> cerbos.policy.v1.TestResults.Summary
-	56, // 79: cerbos.policy.v1.TestResults.Suite.test_cases:type_name -> cerbos.policy.v1.TestResults.TestCase
-	57, // 80: cerbos.policy.v1.TestResults.TestCase.principals:type_name -> cerbos.policy.v1.TestResults.Principal
-	58, // 81: cerbos.policy.v1.TestResults.Principal.resources:type_name -> cerbos.policy.v1.TestResults.Resource
-	59, // 82: cerbos.policy.v1.TestResults.Resource.actions:type_name -> cerbos.policy.v1.TestResults.Action
-	60, // 83: cerbos.policy.v1.TestResults.Action.details:type_name -> cerbos.policy.v1.TestResults.Details
-	0,  // 84: cerbos.policy.v1.TestResults.Details.result:type_name -> cerbos.policy.v1.TestResults.Result
-	62, // 85: cerbos.policy.v1.TestResults.Details.failure:type_name -> cerbos.policy.v1.TestResults.Failure
-	63, // 86: cerbos.policy.v1.TestResults.Details.success:type_name -> cerbos.policy.v1.TestResults.Success
-	75, // 87: cerbos.policy.v1.TestResults.Details.engine_trace:type_name -> cerbos.engine.v1.Trace
-	64, // 88: cerbos.policy.v1.TestResults.OutputFailure.mismatched:type_name -> cerbos.policy.v1.TestResults.OutputFailure.MismatchedValue
-	65, // 89: cerbos.policy.v1.TestResults.OutputFailure.missing:type_name -> cerbos.policy.v1.TestResults.OutputFailure.MissingValue
-	67, // 90: cerbos.policy.v1.TestResults.Failure.expected:type_name -> cerbos.effect.v1.Effect
-	67, // 91: cerbos.policy.v1.TestResults.Failure.actual:type_name -> cerbos.effect.v1.Effect
-	61, // 92: cerbos.policy.v1.TestResults.Failure.outputs:type_name -> cerbos.policy.v1.TestResults.OutputFailure
-	67, // 93: cerbos.policy.v1.TestResults.Success.effect:type_name -> cerbos.effect.v1.Effect
-	74, // 94: cerbos.policy.v1.TestResults.Success.outputs:type_name -> cerbos.engine.v1.OutputEntry
-	70, // 95: cerbos.policy.v1.TestResults.OutputFailure.MismatchedValue.expected:type_name -> google.protobuf.Value
-	70, // 96: cerbos.policy.v1.TestResults.OutputFailure.MismatchedValue.actual:type_name -> google.protobuf.Value
-	70, // 97: cerbos.policy.v1.TestResults.OutputFailure.MissingValue.expected:type_name -> google.protobuf.Value
+	4,  // 0: cerbos.policy.v1.Policy.metadata:type_name -> cerbos.policy.v1.Metadata
+	5,  // 1: cerbos.policy.v1.Policy.resource_policy:type_name -> cerbos.policy.v1.ResourcePolicy
+	9,  // 2: cerbos.policy.v1.Policy.principal_policy:type_name -> cerbos.policy.v1.PrincipalPolicy
+	11, // 3: cerbos.policy.v1.Policy.derived_roles:type_name -> cerbos.policy.v1.DerivedRoles
+	13, // 4: cerbos.policy.v1.Policy.export_variables:type_name -> cerbos.policy.v1.ExportVariables
+	7,  // 5: cerbos.policy.v1.Policy.role_policy:type_name -> cerbos.policy.v1.RolePolicy
+	25, // 6: cerbos.policy.v1.Policy.variables:type_name -> cerbos.policy.v1.Policy.VariablesEntry
+	26, // 7: cerbos.policy.v1.SourceAttributes.attributes:type_name -> cerbos.policy.v1.SourceAttributes.AttributesEntry
+	27, // 8: cerbos.policy.v1.Metadata.annotations:type_name -> cerbos.policy.v1.Metadata.AnnotationsEntry
+	67, // 9: cerbos.policy.v1.Metadata.hash:type_name -> google.protobuf.UInt64Value
+	3,  // 10: cerbos.policy.v1.Metadata.source_attributes:type_name -> cerbos.policy.v1.SourceAttributes
+	6,  // 11: cerbos.policy.v1.ResourcePolicy.rules:type_name -> cerbos.policy.v1.ResourceRule
+	18, // 12: cerbos.policy.v1.ResourcePolicy.schemas:type_name -> cerbos.policy.v1.Schemas
+	14, // 13: cerbos.policy.v1.ResourcePolicy.variables:type_name -> cerbos.policy.v1.Variables
+	15, // 14: cerbos.policy.v1.ResourceRule.condition:type_name -> cerbos.policy.v1.Condition
+	68, // 15: cerbos.policy.v1.ResourceRule.effect:type_name -> cerbos.effect.v1.Effect
+	17, // 16: cerbos.policy.v1.ResourceRule.output:type_name -> cerbos.policy.v1.Output
+	8,  // 17: cerbos.policy.v1.RolePolicy.rules:type_name -> cerbos.policy.v1.RoleRule
+	10, // 18: cerbos.policy.v1.PrincipalPolicy.rules:type_name -> cerbos.policy.v1.PrincipalRule
+	14, // 19: cerbos.policy.v1.PrincipalPolicy.variables:type_name -> cerbos.policy.v1.Variables
+	28, // 20: cerbos.policy.v1.PrincipalRule.actions:type_name -> cerbos.policy.v1.PrincipalRule.Action
+	12, // 21: cerbos.policy.v1.DerivedRoles.definitions:type_name -> cerbos.policy.v1.RoleDef
+	14, // 22: cerbos.policy.v1.DerivedRoles.variables:type_name -> cerbos.policy.v1.Variables
+	15, // 23: cerbos.policy.v1.RoleDef.condition:type_name -> cerbos.policy.v1.Condition
+	29, // 24: cerbos.policy.v1.ExportVariables.definitions:type_name -> cerbos.policy.v1.ExportVariables.DefinitionsEntry
+	30, // 25: cerbos.policy.v1.Variables.local:type_name -> cerbos.policy.v1.Variables.LocalEntry
+	16, // 26: cerbos.policy.v1.Condition.match:type_name -> cerbos.policy.v1.Match
+	31, // 27: cerbos.policy.v1.Match.all:type_name -> cerbos.policy.v1.Match.ExprList
+	31, // 28: cerbos.policy.v1.Match.any:type_name -> cerbos.policy.v1.Match.ExprList
+	31, // 29: cerbos.policy.v1.Match.none:type_name -> cerbos.policy.v1.Match.ExprList
+	32, // 30: cerbos.policy.v1.Output.when:type_name -> cerbos.policy.v1.Output.When
+	34, // 31: cerbos.policy.v1.Schemas.principal_schema:type_name -> cerbos.policy.v1.Schemas.Schema
+	34, // 32: cerbos.policy.v1.Schemas.resource_schema:type_name -> cerbos.policy.v1.Schemas.Schema
+	69, // 33: cerbos.policy.v1.TestOptions.now:type_name -> google.protobuf.Timestamp
+	41, // 34: cerbos.policy.v1.TestOptions.globals:type_name -> cerbos.policy.v1.TestOptions.GlobalsEntry
+	22, // 35: cerbos.policy.v1.TestSuite.tests:type_name -> cerbos.policy.v1.TestTable
+	42, // 36: cerbos.policy.v1.TestSuite.principals:type_name -> cerbos.policy.v1.TestSuite.PrincipalsEntry
+	43, // 37: cerbos.policy.v1.TestSuite.resources:type_name -> cerbos.policy.v1.TestSuite.ResourcesEntry
+	44, // 38: cerbos.policy.v1.TestSuite.aux_data:type_name -> cerbos.policy.v1.TestSuite.AuxDataEntry
+	20, // 39: cerbos.policy.v1.TestSuite.options:type_name -> cerbos.policy.v1.TestOptions
+	45, // 40: cerbos.policy.v1.TestTable.input:type_name -> cerbos.policy.v1.TestTable.Input
+	47, // 41: cerbos.policy.v1.TestTable.expected:type_name -> cerbos.policy.v1.TestTable.Expectation
+	20, // 42: cerbos.policy.v1.TestTable.options:type_name -> cerbos.policy.v1.TestOptions
+	49, // 43: cerbos.policy.v1.Test.name:type_name -> cerbos.policy.v1.Test.TestName
+	70, // 44: cerbos.policy.v1.Test.input:type_name -> cerbos.engine.v1.CheckInput
+	51, // 45: cerbos.policy.v1.Test.expected:type_name -> cerbos.policy.v1.Test.ExpectedEntry
+	20, // 46: cerbos.policy.v1.Test.options:type_name -> cerbos.policy.v1.TestOptions
+	52, // 47: cerbos.policy.v1.Test.expected_outputs:type_name -> cerbos.policy.v1.Test.ExpectedOutputsEntry
+	56, // 48: cerbos.policy.v1.TestResults.suites:type_name -> cerbos.policy.v1.TestResults.Suite
+	55, // 49: cerbos.policy.v1.TestResults.summary:type_name -> cerbos.policy.v1.TestResults.Summary
+	71, // 50: cerbos.policy.v1.SourceAttributes.AttributesEntry.value:type_name -> google.protobuf.Value
+	15, // 51: cerbos.policy.v1.PrincipalRule.Action.condition:type_name -> cerbos.policy.v1.Condition
+	68, // 52: cerbos.policy.v1.PrincipalRule.Action.effect:type_name -> cerbos.effect.v1.Effect
+	17, // 53: cerbos.policy.v1.PrincipalRule.Action.output:type_name -> cerbos.policy.v1.Output
+	16, // 54: cerbos.policy.v1.Match.ExprList.of:type_name -> cerbos.policy.v1.Match
+	33, // 55: cerbos.policy.v1.Schemas.Schema.ignore_when:type_name -> cerbos.policy.v1.Schemas.IgnoreWhen
+	38, // 56: cerbos.policy.v1.TestFixture.Principals.principals:type_name -> cerbos.policy.v1.TestFixture.Principals.PrincipalsEntry
+	39, // 57: cerbos.policy.v1.TestFixture.Resources.resources:type_name -> cerbos.policy.v1.TestFixture.Resources.ResourcesEntry
+	40, // 58: cerbos.policy.v1.TestFixture.AuxData.aux_data:type_name -> cerbos.policy.v1.TestFixture.AuxData.AuxDataEntry
+	72, // 59: cerbos.policy.v1.TestFixture.Principals.PrincipalsEntry.value:type_name -> cerbos.engine.v1.Principal
+	73, // 60: cerbos.policy.v1.TestFixture.Resources.ResourcesEntry.value:type_name -> cerbos.engine.v1.Resource
+	74, // 61: cerbos.policy.v1.TestFixture.AuxData.AuxDataEntry.value:type_name -> cerbos.engine.v1.AuxData
+	71, // 62: cerbos.policy.v1.TestOptions.GlobalsEntry.value:type_name -> google.protobuf.Value
+	72, // 63: cerbos.policy.v1.TestSuite.PrincipalsEntry.value:type_name -> cerbos.engine.v1.Principal
+	73, // 64: cerbos.policy.v1.TestSuite.ResourcesEntry.value:type_name -> cerbos.engine.v1.Resource
+	74, // 65: cerbos.policy.v1.TestSuite.AuxDataEntry.value:type_name -> cerbos.engine.v1.AuxData
+	75, // 66: cerbos.policy.v1.TestTable.OutputExpectations.expected:type_name -> cerbos.engine.v1.OutputEntry
+	48, // 67: cerbos.policy.v1.TestTable.Expectation.actions:type_name -> cerbos.policy.v1.TestTable.Expectation.ActionsEntry
+	46, // 68: cerbos.policy.v1.TestTable.Expectation.outputs:type_name -> cerbos.policy.v1.TestTable.OutputExpectations
+	68, // 69: cerbos.policy.v1.TestTable.Expectation.ActionsEntry.value:type_name -> cerbos.effect.v1.Effect
+	53, // 70: cerbos.policy.v1.Test.OutputEntries.entries:type_name -> cerbos.policy.v1.Test.OutputEntries.EntriesEntry
+	68, // 71: cerbos.policy.v1.Test.ExpectedEntry.value:type_name -> cerbos.effect.v1.Effect
+	50, // 72: cerbos.policy.v1.Test.ExpectedOutputsEntry.value:type_name -> cerbos.policy.v1.Test.OutputEntries
+	71, // 73: cerbos.policy.v1.Test.OutputEntries.EntriesEntry.value:type_name -> google.protobuf.Value
+	1,  // 74: cerbos.policy.v1.TestResults.Tally.result:type_name -> cerbos.policy.v1.TestResults.Result
+	1,  // 75: cerbos.policy.v1.TestResults.Summary.overall_result:type_name -> cerbos.policy.v1.TestResults.Result
+	54, // 76: cerbos.policy.v1.TestResults.Summary.result_counts:type_name -> cerbos.policy.v1.TestResults.Tally
+	58, // 77: cerbos.policy.v1.TestResults.Suite.principals:type_name -> cerbos.policy.v1.TestResults.Principal
+	55, // 78: cerbos.policy.v1.TestResults.Suite.summary:type_name -> cerbos.policy.v1.TestResults.Summary
+	57, // 79: cerbos.policy.v1.TestResults.Suite.test_cases:type_name -> cerbos.policy.v1.TestResults.TestCase
+	58, // 80: cerbos.policy.v1.TestResults.TestCase.principals:type_name -> cerbos.policy.v1.TestResults.Principal
+	59, // 81: cerbos.policy.v1.TestResults.Principal.resources:type_name -> cerbos.policy.v1.TestResults.Resource
+	60, // 82: cerbos.policy.v1.TestResults.Resource.actions:type_name -> cerbos.policy.v1.TestResults.Action
+	61, // 83: cerbos.policy.v1.TestResults.Action.details:type_name -> cerbos.policy.v1.TestResults.Details
+	1,  // 84: cerbos.policy.v1.TestResults.Details.result:type_name -> cerbos.policy.v1.TestResults.Result
+	63, // 85: cerbos.policy.v1.TestResults.Details.failure:type_name -> cerbos.policy.v1.TestResults.Failure
+	64, // 86: cerbos.policy.v1.TestResults.Details.success:type_name -> cerbos.policy.v1.TestResults.Success
+	76, // 87: cerbos.policy.v1.TestResults.Details.engine_trace:type_name -> cerbos.engine.v1.Trace
+	65, // 88: cerbos.policy.v1.TestResults.OutputFailure.mismatched:type_name -> cerbos.policy.v1.TestResults.OutputFailure.MismatchedValue
+	66, // 89: cerbos.policy.v1.TestResults.OutputFailure.missing:type_name -> cerbos.policy.v1.TestResults.OutputFailure.MissingValue
+	68, // 90: cerbos.policy.v1.TestResults.Failure.expected:type_name -> cerbos.effect.v1.Effect
+	68, // 91: cerbos.policy.v1.TestResults.Failure.actual:type_name -> cerbos.effect.v1.Effect
+	62, // 92: cerbos.policy.v1.TestResults.Failure.outputs:type_name -> cerbos.policy.v1.TestResults.OutputFailure
+	68, // 93: cerbos.policy.v1.TestResults.Success.effect:type_name -> cerbos.effect.v1.Effect
+	75, // 94: cerbos.policy.v1.TestResults.Success.outputs:type_name -> cerbos.engine.v1.OutputEntry
+	71, // 95: cerbos.policy.v1.TestResults.OutputFailure.MismatchedValue.expected:type_name -> google.protobuf.Value
+	71, // 96: cerbos.policy.v1.TestResults.OutputFailure.MismatchedValue.actual:type_name -> google.protobuf.Value
+	71, // 97: cerbos.policy.v1.TestResults.OutputFailure.MissingValue.expected:type_name -> google.protobuf.Value
 	98, // [98:98] is the sub-list for method output_type
 	98, // [98:98] is the sub-list for method input_type
 	98, // [98:98] is the sub-list for extension type_name
@@ -5014,7 +5077,7 @@ func file_cerbos_policy_v1_policy_proto_init() {
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_cerbos_policy_v1_policy_proto_rawDesc,
-			NumEnums:      1,
+			NumEnums:      2,
 			NumMessages:   65,
 			NumExtensions: 0,
 			NumServices:   0,
