@@ -7,18 +7,20 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"strconv"
 	"sync"
 	"time"
+
+	badgerv4 "github.com/dgraph-io/badger/v4"
+	"github.com/sourcegraph/conc/pool"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
 
 	auditv1 "github.com/cerbos/cerbos/api/genpb/cerbos/audit/v1"
 	"github.com/cerbos/cerbos/internal/audit"
 	"github.com/cerbos/cerbos/internal/audit/local"
 	"github.com/cerbos/cerbos/internal/config"
 	logsv1 "github.com/cerbos/cloud-api/genpb/cerbos/cloud/logs/v1"
-	badgerv4 "github.com/dgraph-io/badger/v4"
-	"github.com/dgraph-io/ristretto/z"
-	"github.com/sourcegraph/conc/pool"
-	"go.uber.org/zap"
 )
 
 const (
@@ -40,7 +42,7 @@ func init() {
 
 		logger := zap.L().Named("auditlog").With(zap.String("backend", Backend))
 
-		syncer, err := NewIngestSyncer(&conf.Ingest, logger)
+		syncer, err := NewIngestSyncer(logger)
 		if err != nil {
 			return nil, err
 		}
@@ -52,11 +54,14 @@ func init() {
 type Log struct {
 	syncer IngestSyncer
 	*local.Log
-	logger                         *zap.Logger
-	filter                         *AuditLogFilter
-	muTimer                        *mutexTimer
-	minFlushInterval, flushTimeout time.Duration
-	maxBatchSize, numGo            int
+	logger           *zap.Logger
+	filter           *AuditLogFilter
+	pool             *pool.ContextPool
+	cancel           context.CancelFunc
+	minFlushInterval time.Duration
+	flushTimeout     time.Duration
+	maxBatchSize     int
+	numGo            int
 }
 
 func NewLog(conf *Conf, decisionFilter audit.DecisionLogEntryFilter, syncer IngestSyncer, logger *zap.Logger) (*Log, error) {
@@ -77,6 +82,8 @@ func NewLog(conf *Conf, decisionFilter audit.DecisionLogEntryFilter, syncer Inge
 		return nil, err
 	}
 
+	ctx, cancelFn := context.WithCancel(context.Background())
+
 	log := &Log{
 		Log:              localLog,
 		syncer:           syncer,
@@ -86,13 +93,11 @@ func NewLog(conf *Conf, decisionFilter audit.DecisionLogEntryFilter, syncer Inge
 		flushTimeout:     flushTimeout,
 		maxBatchSize:     maxBatchSize,
 		numGo:            numGo,
-		muTimer:          newMutexTimer(),
+		cancel:           cancelFn,
+		pool:             pool.New().WithContext(ctx),
 	}
 
-	localLog.RegisterCallback(func(waitCh chan<- struct{}) {
-		log.schedule(waitCh)
-	})
-
+	log.pool.Go(log.syncLoop)
 	return log, nil
 }
 
@@ -144,65 +149,35 @@ func (l *Log) WriteDecisionLogEntry(ctx context.Context, record audit.DecisionLo
 	return l.Write(ctx, key, value)
 }
 
-type mutexTimer struct {
-	expireCh chan struct{}
-	t        *time.Timer
-	mu       sync.RWMutex
-}
-
-func newMutexTimer() *mutexTimer {
-	return &mutexTimer{
-		expireCh: make(chan struct{}, 1),
-	}
-}
-
-func (mt *mutexTimer) set(waitDuration time.Duration) {
-	mt.mu.Lock()
-	defer mt.mu.Unlock()
-
-	if mt.t == nil {
-		mt.t = time.AfterFunc(waitDuration, func() {
-			mt.expireCh <- struct{}{}
-
-			mt.mu.Lock()
-			mt.t = nil
-			mt.mu.Unlock()
-		})
-	}
-}
-
-func (mt *mutexTimer) wait() {
-	{
-		mt.mu.RLock()
-		defer mt.mu.RUnlock()
-		if mt.t == nil {
-			return
+func (l *Log) syncLoop(ctx context.Context) error {
+	ticker := time.NewTicker(l.minFlushInterval)
+	for {
+		select {
+		case <-ticker.C:
+			ticker.Stop()
+			delay := l.schedule()
+			ticker.Reset(delay)
+		case <-ctx.Done():
+			return nil
 		}
 	}
-
-	<-mt.expireCh
 }
 
-func (l *Log) schedule(waitCh chan<- struct{}) {
-	l.muTimer.wait()
-
+func (l *Log) schedule() time.Duration {
+	l.logger.Log(zapcore.Level(-3), "Scheduling stream")
 	if err := l.streamLogs(); err != nil {
 		var ingestErr ErrIngestBackoff
 		if errors.As(err, &ingestErr) {
 			l.logger.Warn("svc-ingest issued backoff", zap.Error(err))
-			l.muTimer.set(ingestErr.Backoff)
-			go l.schedule(waitCh)
-			return
+			if ingestErr.Backoff < l.minFlushInterval {
+				return l.minFlushInterval
+			}
+			return ingestErr.Backoff
 		}
-		l.logger.Warn("Failed sync", zap.Error(err))
+		l.logger.Error("Audit log sync failed", zap.Error(err))
 	}
 
-	// Set a min wait duration regardless of if events are pending.
-	// This prevents a retry completion occurring immediately before the next sync
-	// (and therefore burdening the backend).
-	l.muTimer.set(l.minFlushInterval)
-
-	waitCh <- struct{}{}
+	return l.minFlushInterval
 }
 
 func (l *Log) streamLogs() error {
@@ -210,61 +185,106 @@ func (l *Log) streamLogs() error {
 	// avoid the penalty of per-key string inspection when inferring the type down the line.
 	ctx := context.Background()
 
-	p := pool.New().WithContext(ctx).WithCancelOnError().WithFirstError()
+	p := pool.New().WithContext(ctx).WithCancelOnError()
 	p.Go(func(ctx context.Context) error {
-		return l.streamPrefix(ctx, logsv1.IngestBatch_ENTRY_KIND_ACCESS_LOG, AccessSyncPrefix)
-	})
-	p.Go(func(ctx context.Context) error {
-		return l.streamPrefix(ctx, logsv1.IngestBatch_ENTRY_KIND_DECISION_LOG, DecisionSyncPrefix)
+		l.logger.Log(zapcore.Level(-2), "Streaming access logs")
+		if err := l.streamPrefix(ctx, logsv1.IngestBatch_ENTRY_KIND_ACCESS_LOG, AccessSyncPrefix); err != nil {
+			l.logger.Warn("Failed to stream access logs", zap.Error(err))
+			return fmt.Errorf("failed to stream access logs: %w", err)
+		}
+		l.logger.Log(zapcore.Level(-2), "Finished streaming access logs")
+		return nil
 	})
 
-	return p.Wait()
+	p.Go(func(ctx context.Context) error {
+		l.logger.Log(zapcore.Level(-2), "Streaming decision logs")
+		if err := l.streamPrefix(ctx, logsv1.IngestBatch_ENTRY_KIND_DECISION_LOG, DecisionSyncPrefix); err != nil {
+			l.logger.Warn("Failed to stream decision logs", zap.Error(err))
+			return fmt.Errorf("failed to stream decision logs: %w", err)
+		}
+		l.logger.Log(zapcore.Level(-2), "Finished streaming decision logs")
+		return nil
+	})
+
+	if err := p.Wait(); err != nil {
+		l.logger.Warn("Failed to stream logs", zap.Error(err))
+		return fmt.Errorf("failed to stream logs: %w", err)
+	}
+	return nil
 }
 
+var keysPool = &sync.Pool{}
+
 func (l *Log) streamPrefix(ctx context.Context, kind logsv1.IngestBatch_EntryKind, prefix []byte) error {
-	// BadgerDB transactions work with snapshot isolation so we only take a view of the DB.
-	// Subsequent writes aren't blocked.
-	stream := l.Db.NewStream()
-	stream.NumGo = l.numGo
-	stream.Prefix = prefix
+	logger := l.logger.With(zap.Stringer("kind", kind))
 
-	stream.Send = func(buf *z.Buffer) error {
-		kvList, err := badgerv4.BufferToKVList(buf)
-		if err != nil {
-			return err
-		}
-
-		keys := make([][]byte, len(kvList.Kv))
-		for i, kv := range kvList.Kv {
-			keys[i] = kv.Key
-		}
-
-		for i := 0; i < len(keys); i += l.maxBatchSize {
-			end := i + l.maxBatchSize
-			if end > len(keys) {
-				end = len(keys)
-			}
-
-			if err := l.syncThenDelete(ctx, kind, keys[i:end]); err != nil {
-				return err
-			}
+	syncKeys := func(keys [][]byte) error {
+		logger.Log(zapcore.Level(-3), "Syncing and deleting batch")
+		if err := l.syncThenDelete(ctx, kind, keys); err != nil {
+			return fmt.Errorf("failed to sync and delete logs: %w", err)
 		}
 
 		return nil
 	}
 
-	return stream.Orchestrate(ctx)
+	opts := badgerv4.DefaultIteratorOptions
+	opts.Prefix = prefix
+	opts.PrefetchValues = false
+	return l.Db.View(func(txn *badgerv4.Txn) error {
+		it := txn.NewIterator(opts)
+		defer it.Close()
+
+		var keys [][]byte
+		if keysIface := keysPool.Get(); keysIface == nil {
+			keys = make([][]byte, l.maxBatchSize)
+		} else {
+			keys = *(keysIface.(*[][]byte)) //nolint:forcetypeassert
+		}
+		defer keysPool.Put(&keys)
+
+		var i int
+		for it.Seek(prefix); it.ValidForPrefix(prefix); it.Next() {
+			item := it.Item()
+
+			if keys[i] == nil {
+				keys[i] = make([]byte, len(item.Key()))
+			} else {
+				clear(keys[i])
+			}
+			copy(keys[i], item.Key())
+
+			i++
+			if i == l.maxBatchSize {
+				if err := syncKeys(keys); err != nil {
+					return err
+				}
+
+				i = 0
+			}
+		}
+
+		return syncKeys(keys[:i])
+	})
 }
 
 func (l *Log) syncThenDelete(ctx context.Context, kind logsv1.IngestBatch_EntryKind, syncKeys [][]byte) error {
+	logger := l.logger.With(zap.Stringer("kind", kind))
+	logger.Log(zapcore.Level(-3), "Getting ingest batch entries")
 	entries, err := l.getIngestBatchEntries(syncKeys, kind)
 	if err != nil {
-		return err
+		logger.Log(zapcore.Level(-2), "Failed to get ingest batch entries", zap.Error(err))
+		return fmt.Errorf("failed to get ingest batch entries: %w", err)
 	}
 
+	if len(entries) == 0 {
+		logger.Log(zapcore.Level(-3), "Ingest batch is empty")
+		return nil
+	}
+
+	logger.Log(zapcore.Level(-3), "Generating audit ID")
 	batchID, err := audit.NewID()
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to generate audit ID: %w", err)
 	}
 
 	{
@@ -276,41 +296,46 @@ func (l *Log) syncThenDelete(ctx context.Context, kind logsv1.IngestBatch_EntryK
 			Entries: entries,
 		}
 
-		if err := l.filter.Filter(&logsv1.IngestBatch{
-			Id:      string(batchID),
-			Entries: entries,
-		}); err != nil {
-			return err
+		logger.Log(zapcore.Level(-3), "Filtering batch of "+strconv.Itoa(len(ingestBatch.Entries)))
+		if err := l.filter.Filter(ingestBatch); err != nil {
+			return fmt.Errorf("failed to filter batch: %w", err)
 		}
 
+		logger.Log(zapcore.Level(-3), "Syncing batch of "+strconv.Itoa(len(ingestBatch.Entries)))
 		if err := l.syncer.Sync(ctx, ingestBatch); err != nil {
-			return err
+			return fmt.Errorf("failed to sync batch: %w", err)
 		}
 	}
 
 	wb := l.Db.NewWriteBatch()
 	defer wb.Cancel()
 
+	logger.Log(zapcore.Level(-3), "Deleting synced keys")
 	for _, k := range syncKeys {
 		if err := wb.Delete(k); err != nil {
 			if errors.Is(err, badgerv4.ErrDiscardedTxn) {
+				wb.Cancel()
 				wb = l.Db.NewWriteBatch()
 				_ = wb.Delete(k)
 			} else {
-				return err
+				return fmt.Errorf("failed to delete key: %w", err)
 			}
 		}
 	}
 
+	logger.Log(zapcore.Level(-3), "Flushing write batch")
 	return wb.Flush()
 }
 
 func (l *Log) getIngestBatchEntries(syncKeys [][]byte, kind logsv1.IngestBatch_EntryKind) ([]*logsv1.IngestBatch_Entry, error) {
-	entries := make([]*logsv1.IngestBatch_Entry, len(syncKeys))
+	entries := make([]*logsv1.IngestBatch_Entry, 0, len(syncKeys))
 	if err := l.Db.Update(func(txn *badgerv4.Txn) error {
-		for i, k := range syncKeys {
+		for _, k := range syncKeys {
 			syncItem, err := txn.Get(k)
 			if err != nil {
+				if errors.Is(err, badgerv4.ErrKeyNotFound) {
+					continue
+				}
 				return err
 			}
 
@@ -324,6 +349,9 @@ func (l *Log) getIngestBatchEntries(syncKeys [][]byte, kind logsv1.IngestBatch_E
 
 			logItem, err := txn.Get(logKey)
 			if err != nil {
+				if errors.Is(err, badgerv4.ErrKeyNotFound) {
+					continue
+				}
 				return err
 			}
 
@@ -365,7 +393,7 @@ func (l *Log) getIngestBatchEntries(syncKeys [][]byte, kind logsv1.IngestBatch_E
 				return err
 			}
 
-			entries[i] = entry
+			entries = append(entries, entry)
 		}
 
 		return nil
@@ -378,4 +406,12 @@ func (l *Log) getIngestBatchEntries(syncKeys [][]byte, kind logsv1.IngestBatch_E
 
 func (l *Log) Backend() string {
 	return Backend
+}
+
+func (l *Log) Close() (outErr error) {
+	l.cancel()
+	outErr = errors.Join(outErr, l.pool.Wait())
+	outErr = errors.Join(outErr, l.Db.Sync())
+	outErr = errors.Join(outErr, l.Log.Close())
+	return outErr
 }
