@@ -5,12 +5,10 @@ package hub
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"errors"
 	"fmt"
 	"io"
-	"os"
+	"regexp"
 	"strconv"
 	"sync"
 	"time"
@@ -22,11 +20,9 @@ import (
 	"github.com/cerbos/cerbos/internal/namer"
 	"github.com/cerbos/cerbos/internal/observability/metrics"
 	"github.com/cerbos/cerbos/internal/storage"
-	"github.com/cerbos/cerbos/internal/util"
 	"github.com/cerbos/cloud-api/base"
 	cloudapi "github.com/cerbos/cloud-api/bundle"
 	"github.com/cerbos/cloud-api/credentials"
-	"github.com/go-logr/zapr"
 	"github.com/spf13/afero"
 	"go.uber.org/zap"
 )
@@ -41,6 +37,8 @@ const (
 var (
 	_ storage.BinaryStore = (*RemoteSource)(nil)
 	_ storage.Reloadable  = (*RemoteSource)(nil)
+
+	playgroundLabelPattern = regexp.MustCompile(`^playground/[A-Z0-9]{12}$`)
 )
 
 type CloudAPIClient interface {
@@ -48,71 +46,41 @@ type CloudAPIClient interface {
 	GetBundle(context.Context, string) (string, error)
 	WatchBundle(context.Context, string) (cloudapi.WatchHandle, error)
 	GetCachedBundle(string) (string, error)
+	HubCredentials() *credentials.Credentials
 }
 
 // RemoteSource implements a bundle store that loads bundles from a remote source.
 type RemoteSource struct {
-	credentials *credentials.Credentials
-	log         *zap.Logger
-	conf        *Conf
-	bundle      *Bundle
-	scratchFS   afero.Fs
-	client      CloudAPIClient
-	mu          sync.RWMutex
-	healthy     bool
+	log        *zap.Logger
+	conf       *Conf
+	bundle     *Bundle
+	scratchFS  afero.Fs
+	client     CloudAPIClient
+	mu         sync.RWMutex
+	healthy    bool
+	playground bool
 }
 
 func NewRemoteSource(conf *Conf) (*RemoteSource, error) {
-	credentials, err := conf.Credentials.ToCredentials()
-	if err != nil {
-		return nil, fmt.Errorf("invalid credentials: %w", err)
-	}
-
-	logger := zap.L().Named(DriverName).With(zap.String("label", conf.Remote.BundleLabel))
-	scratchFS := afero.NewBasePathFs(afero.NewOsFs(), conf.Remote.TempDir)
-	return &RemoteSource{log: logger, conf: conf, healthy: false, scratchFS: scratchFS, credentials: credentials}, nil
+	return &RemoteSource{
+		conf:       conf,
+		healthy:    false,
+		playground: playgroundLabelPattern.MatchString(conf.Remote.BundleLabel),
+		log:        zap.L().Named(DriverName).With(zap.String("label", conf.Remote.BundleLabel)),
+		scratchFS:  afero.NewBasePathFs(afero.NewOsFs(), conf.Remote.TempDir),
+	}, nil
 }
 
 func (s *RemoteSource) Init(ctx context.Context) error {
-	pdpID := util.PDPIdentifier(s.conf.Credentials.PDPID)
-	s.log = s.log.With(zap.String("instance", pdpID.Instance))
-
-	tlsConf := &tls.Config{
-		MinVersion: tls.VersionTLS13,
-		ServerName: s.conf.Remote.Connection.TLS.Authority,
+	hubInstance, err := hub.Get()
+	if err != nil {
+		return fmt.Errorf("failed to establish Cerbos Hub connection: %w", err)
 	}
 
-	caCertPath := s.conf.Remote.Connection.TLS.CACert
-	if caCertPath != "" {
-		caCert, err := os.ReadFile(caCertPath)
-		if err != nil {
-			return fmt.Errorf("failed to read CA cert from %q: %w", caCertPath, err)
-		}
-
-		tlsConf.RootCAs = x509.NewCertPool()
-		if !tlsConf.RootCAs.AppendCertsFromPEM(caCert) {
-			return fmt.Errorf("failed to parse CA certs")
-		}
-	}
-
-	clientConf := cloudapi.ClientConf{
-		ClientConf: base.ClientConf{
-			Logger:            zapr.NewLogger(s.log),
-			PDPIdentifier:     pdpID,
-			TLS:               tlsConf,
-			Credentials:       s.credentials,
-			APIEndpoint:       s.conf.Remote.Connection.APIEndpoint,
-			BootstrapEndpoint: s.conf.Remote.Connection.BootstrapEndpoint,
-			RetryWaitMin:      s.conf.Remote.Connection.MinRetryWait,
-			RetryWaitMax:      s.conf.Remote.Connection.MaxRetryWait,
-			RetryMaxAttempts:  int(s.conf.Remote.Connection.NumRetries),
-			HeartbeatInterval: s.conf.Remote.Connection.HeartbeatInterval,
-		},
+	client, err := hubInstance.BundleClient(cloudapi.ClientConf{
 		CacheDir: s.conf.Remote.CacheDir,
 		TempDir:  s.conf.Remote.TempDir,
-	}
-
-	client, err := cloudapi.NewClient(clientConf)
+	})
 	if err != nil {
 		return fmt.Errorf("failed to create API client: %w", err)
 	}
@@ -164,16 +132,19 @@ func shouldWorkOffline() bool {
 }
 
 func (s *RemoteSource) fetchBundle(ctx context.Context) error {
-	s.log.Info("Fetching bundle")
-	bdlPath, err := s.client.BootstrapBundle(ctx, s.conf.Remote.BundleLabel)
-	if err == nil {
-		s.log.Debug("Using bootstrap bundle")
-		return s.swapBundle(bdlPath)
+	if !s.playground {
+		s.log.Info("Fetching bootstrap bundle")
+		bdlPath, err := s.client.BootstrapBundle(ctx, s.conf.Remote.BundleLabel)
+		if err == nil {
+			s.log.Debug("Using bootstrap bundle")
+			return s.swapBundle(bdlPath)
+		}
+
+		s.log.Warn("Failed to fetch bootstrap bundle", zap.Error(err))
 	}
 
-	s.log.Warn("Failed to fetch bootstrap bundle", zap.Error(err))
-	s.log.Info("Attempting to fetch bundle from the API")
-	bdlPath, err = s.client.GetBundle(ctx, s.conf.Remote.BundleLabel)
+	s.log.Info("Fetching bundle from the API")
+	bdlPath, err := s.client.GetBundle(ctx, s.conf.Remote.BundleLabel)
 	if err != nil {
 		s.log.Error("Failed to fetch bundle using the API", zap.Error(err))
 		metrics.Inc(ctx, metrics.BundleFetchErrorsCount())
@@ -212,13 +183,18 @@ func (s *RemoteSource) removeBundle(healthy bool) {
 func (s *RemoteSource) swapBundle(bundlePath string) error {
 	s.log.Debug("Swapping bundle", zap.String("path", bundlePath))
 
-	bundle, err := Open(OpenOpts{
-		Source:      "remote",
-		BundlePath:  bundlePath,
-		ScratchFS:   s.scratchFS,
-		Credentials: s.credentials,
-		CacheSize:   s.conf.CacheSize,
-	})
+	opts := OpenOpts{
+		Source:     "remote",
+		BundlePath: bundlePath,
+		ScratchFS:  s.scratchFS,
+		CacheSize:  s.conf.CacheSize,
+	}
+
+	if !s.playground {
+		opts.Credentials = s.client.HubCredentials()
+	}
+
+	bundle, err := Open(opts)
 	if err != nil {
 		s.log.Error("Failed to open bundle", zap.Error(err))
 		return fmt.Errorf("failed to open bundle: %w", err)
